Unify kms provider config for reusability
